requirements.txt
click==7.1.2
Flask==1.1.4
importlib-metadata==5.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.18.0
packaging==21.3
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.21.0
Werkzeug==1.0.1
zipp==3.9.0
